Use proper includes and remove anonymous namespace
Anonymous namespace doesn't work with UE4 Unity builds. Everything can
see it so it was quite hard to follow where the constants are coming
from. Now the constants are passed around instead of used from the
namespace, and the namespace now has an explicit name.

Also tried to fix some includes, which are also seen everywhere with
Unity builds. At the same time switch to UE4 math from cmath.