Almost a complete rewrite of non-GRPC code
Now the results are collected by Speechly plugin, and exposed as
delegates whenever the data changes. So the raw message parsing is now
handled by Speechly, which makes the client code a lot simpler, but also
a little bit less flexible as not everything is exposed.

But those features can still be exposed if necessary in future updates.

Audio code also has some changes, which I didn't separate form this
commit as it's quite intermingled with the other changes. Changed the
way buffering is done, to reduce the possibility of sending too small
buffers on start.